pyproject.toml
[tool.poetry]
name = "tokamak2"
description = "Interference generator"
authors = ["Kevin Zeidler"]
version = "0.1.0"
packages = [
    { include = "tokamak", from = "src" },
    { include = "interference", from = "src" }
    
]
[tool.poetry.dependencies]
python = ">=3.6"
codegen = ">=1.0"
dash-bootstrap-components = ">=0.7"
dash-core-components = ">=1.5"
dash-daq = ">=0.3.0"
dash-html-components = ">=1.0"
flask-caching = ">=1.7.2"
gemeinsprache = ">=0.0"
redis = ">=3.3.11"
termcolor = ">=1.1.0"
werkzeug = "*"
psutil = "^5.6"
requests = "^2.22"
orca = "^1.5"
lxml = "^4.4"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]
ptpython = ">2.0"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
